Add conversation to bot with /start command
Implemented start of conversation with bot with /start command.
Bot requests user to send a YouTube video or playlist URL.
Allow user to finish conversation with /cancel command.
ConversationHandler used because it allows for a more complex flow of
messages between the bot and the user, in states or steps.